feat: slog wrapper
